completed ci build

+++ service/routes.py
@@ -4,7 +4,7 @@ Account Service
 This microservice handles the lifecycle of Accounts
 """
 # pylint: disable=unused-import
-from flask import jsonify, request, make_response, abort, url_for   # noqa; F401
+from flask import jsonify, request, make_response, abort, url_for  # noqa: F401
 from service.models import Account
 from service.common import status  # HTTP Status Codes
 from . import app  # Import Flask application
@@ -22,6 +22,8 @@ def health():
 ######################################################################
 # GET INDEX
 ######################################################################
+
+
 @app.route("/")
 def index():
     """Root URL response"""
@@ -38,6 +40,8 @@ def index():
 ######################################################################
 # CREATE A NEW ACCOUNT
 ######################################################################
+
+
 @app.route("/accounts", methods=["POST"])
 def create_accounts():
     """
@@ -46,50 +50,59 @@ def create_accounts():
     """
     app.logger.info("Request to create an Account")
     check_content_type("application/json")
+
     account = Account()
     account.deserialize(request.get_json())
     account.create()
+
     message = account.serialize()
+
     # Uncomment once get_accounts has been implemented
     # location_url = url_for("get_accounts", account_id=account.id, _external=True)
     location_url = "/"  # Remove once get_accounts has been implemented
+
     return make_response(
-        jsonify(message), status.HTTP_201_CREATED, {"Location": location_url}
+        jsonify(message),
+        status.HTTP_201_CREATED,
+        {"Location": location_url},
     )
+
 
 ######################################################################
 # LIST ALL ACCOUNTS
 ######################################################################
 
-# ... place you code here to LIST accounts ...
-@app.route('/accounts',  methods=["GET"])
-def list_all_account():
-    """List All Account"""
-    
-   
-    app.logger.info("Request to list Accounts")
-    accounts = Account.all()
 
+@app.route("/accounts", methods=["GET"])
+def list_all_account():
+    """List All Accounts"""
+    app.logger.info("Request to list Accounts")
+
+    accounts = Account.all()
     account_list = [account.serialize() for account in accounts]
 
     app.logger.info("Returning [%s] accounts", len(account_list))
     return jsonify(account_list), status.HTTP_200_OK
 
+
 ######################################################################
 # READ AN ACCOUNT
 ######################################################################
 
-# ... place you code here to READ an account ...
+
 @app.route("/accounts/<int:account_id>", methods=["GET"])
 def read_account(account_id):
     """
-    Read account
+    Read an Account
     """
     app.logger.info("Request to read an Account with id: %s", account_id)
 
     account = Account.find(account_id)
     if not account:
-        abort(status.HTTP_404_NOT_FOUND, f"Account with id [{account_id}] could not be found.")
+        abort(
+            status.HTTP_404_NOT_FOUND,
+            f"Account with id [{account_id}] could not be found.",
+        )
 
     return account.serialize(), status.HTTP_200_OK
 
@@ -98,16 +111,22 @@ def read_account(account_id):
 # UPDATE AN EXISTING ACCOUNT
 ######################################################################
 
-@app.route('/accounts/<int:account_id>', methods = ["PUT"])
-def update_account(account_id):
 
+@app.route("/accounts/<int:account_id>", methods=["PUT"])
+def update_account(account_id):
     """
     Update an Account
     This endpoint will update an Account based on the posted data
     """
     app.logger.info("Request to update an Account with id: %s", account_id)
 
-    account = Account.find(account_id)  
+    account = Account.find(account_id)
+    if not account:
+        abort(
+            status.HTTP_404_NOT_FOUND,
+            f"Account with id [{account_id}] could not be found.",
+        )
+
     account.deserialize(request.get_json())
     account.update()
 
@@ -118,15 +137,22 @@ def update_account(account_id):
 # DELETE AN ACCOUNT
 ######################################################################
 
-@app.route('/accounts/<int:account_id>', methods = ["DELETE"])
-def delete_account(account_id):
 
+@app.route("/accounts/<int:account_id>", methods=["DELETE"])
+def delete_account(account_id):
     """
     Delete an Account
     This endpoint will delete an Account based on the posted data
     """
     app.logger.info("Request to delete an Account with id: %s", account_id)
+
     account = Account.find(account_id)
+    if not account:
+        abort(
+            status.HTTP_404_NOT_FOUND,
+            f"Account with id [{account_id}] could not be found.",
+        )
+
     account.delete()
 
     return "", status.HTTP_204_NO_CONTENT
@@ -140,9 +166,12 @@ def delete_account(account_id):
 def check_content_type(media_type):
     """Checks that the media type is correct"""
     content_type = request.headers.get("Content-Type")
+
     if content_type and content_type == media_type:
         return
+
     app.logger.error("Invalid Content-Type: %s", content_type)
+
     abort(
         status.HTTP_415_UNSUPPORTED_MEDIA_TYPE,
         f"Content-Type must be {media_type}",
